infrastructure/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.51.1
aws-cdk.asset-awscli-v1==2.2.17
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.23
blessed==1.19.1
boto3==1.26.16
botocore==1.29.16
cattrs==22.2.0
chalice==1.27.3
click==8.1.3
constructs==10.1.169
exceptiongroup==1.0.4
flake8==6.0.0
inquirer==2.10.1
jmespath==1.0.1
jsii==1.71.0
mccabe==0.7.0
publication==0.0.3
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.3
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
